+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to the changelog-helper will be documented in this file.
 
 ## [unreleased]
 
+## [1.2.0] - 2025-09-19
+
 ### Added
 
 - Release command now accepts specific version numbers (e.g., 3.2.1) in addition to release levels (major, minor, patch)
+- Added new release build
 
 
 ### Changed